+++ whelk-core/src/main/resources/romanizer/mn-lessing.txt
